refactor: combine all hook definitions into one module
