+++ virtool_workflow/hooks.py
@@ -4,11 +4,53 @@ Hooks provide a way to do things when events happen during the workflow lifecycl
 import asyncio
 from concurrent import futures
 
-from virtool_workflow.execution.hooks.fixture_hooks import FixtureHook
-from virtool_workflow.execution.hooks.hooks import Hook
-from virtool_workflow.execution.hooks.workflow_hooks import *
+from virtool_workflow.execution.hooks import Hook
 
-on_success = FixtureHook("on_success")
+on_result = Hook("on_result")
+"""
+Triggered when a workflow has completed and a result is available.
+
+.. code-block:: python
+
+    @on_result
+    async def use_result(workflow: Workflow, results: Dict[str, Any]):
+        ...
+
+This Hook is triggered before the result of the workflow is stored. As
+such the result can be mutated within the callback and that change will be
+reflected in the final result.
+"""
+
+on_step_start = Hook("on_step_start")
+"""
+Triggered before each workflow step is executed.
+
+The :class:`WorkflowStep` object is available via the `current_step` fixture.
+
+.. code_block:: python
+
+    @on_step_start
+    async def use_step(current_step):
+        ...
+"""
+
+on_step_finish = Hook("on_step_end")
+"""
+Triggered after each workflow step is executed.
+
+The :class:`WorkflowStep` object is available via the `current_step` fixture.
+
+@on_step_finish
+async def use_step(current_step):
+    ...
+"""
+
+on_workflow_start = Hook("on_workflow_start")
+"""
+Triggered at the start of the workflow, before any steps are executed.
+"""
+
+on_success = Hook("on_success")
 """
 Triggered when a job completes successfully.
 
@@ -21,7 +63,7 @@ Parameters supplied are the `Workflow` instance and the results dict.
         ...
 """
 
-on_failure = FixtureHook("on_failure")
+on_failure = Hook("on_failure")
 """
 Triggered when a job fails to complete. The exception
 which caused the failure will be found in the `error` fixture.
@@ -33,7 +75,7 @@ which caused the failure will be found in the `error` fixture.
         ...
 """
 
-on_finish = FixtureHook("on_finish")
+on_finish = Hook("on_finish")
 """
 Triggered when a job finishes, regardless of success or failure.
 
@@ -45,7 +87,7 @@ Triggered when a job finishes, regardless of success or failure.
 """
 
 
-on_finalize = FixtureHook("on_finalize")
+on_finalize = Hook("on_finalize")
 """
 Triggered after job finishes, regardless of success or failure.
 
@@ -53,7 +95,7 @@ Intended for finalization actions such as closing the fixture scope.
 """
 
 
-on_cancelled = FixtureHook("on_cancelled")
+on_cancelled = Hook("on_cancelled")
 """
 Triggered when a job is cancelled.
 
@@ -71,7 +113,7 @@ async def _trigger_on_cancelled(error: Exception, scope):
         await on_cancelled.trigger(scope, error)
 
 
-on_load_config = FixtureHook("on_load_config")
+on_load_config = Hook("on_load_config")
 """
 Triggered after the config is loaded from the CLI arguments and environment variables. A SimpleNamespace object
 is provided which has an attribute (sharing the same name as the fixture) for each configuration fixture in
@@ -85,7 +127,7 @@ is provided which has an attribute (sharing the same name as the fixture) for ea
             ...
 """
 
-before_result_upload = FixtureHook("before_result_upload")
+before_result_upload = Hook("before_result_upload")
 """Triggered after the result is ready to be uploaded, but before it is actually uploaded."""
 
 __all__ = [
